server/pyproject.toml
[project]
name = "demorph-server"
version = "0.1.0"
description = "DeMorph is a multimodal deepfake detection solution."
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Akshat Mishra", email="akshatdmishra2005@gmail.com" },
]
classifiers = [
    "Private :: Do Not Upload",
]
dependencies = [
    "alembic>=1.16.2",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.115.13",
    "loguru>=0.7.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.0",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "poethepoet>=0.38.0",
    "pre-commit>=4.5.0",
    "ruff>=0.14.9",
]

[tool.poe.tasks]
start-dev = "uvicorn src.server:api --reload"
start = "uvicorn src.server:api --host 0.0.0.0 --port ${PORT:-8000}"

lint = "ruff check . --fix"
format = "ruff format ."
fix = ["lint", "format"]

test = "pytest"
test-cov = "pytest --cov=src"
type-check = "mypy src"
audit = "uv audit"

[tool.poe.tasks.db-migrate]
cmd = "alembic revision --autogenerate -m \"${message}\""
help = "Create a new autogenerated migration file."
args = [
    { name = "message", options = ["-m", "--message"], help = "Migration message." }
]

[tool.poe.tasks.db-upgrade]
cmd = "alembic upgrade head"
help = "Apply all pending migrations to database."

[tool.poe.tasks.db-downgrade]
cmd = "alembic downgrade -1"
help = "Revert the last migration."

[tool.pytest.ini_options]
testpaths = "tests"
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test* *Tests"
addopts = "-vv -x"
asyncio_mode = "auto"
pythonpath = ["."]

[tool.mypy]
check_untyped_defs = true
namespace_packages = true
plugins = "sqlalchemy.ext.mypy.plugin"
pretty = true
show_error_context = true
exclude = ".venv"

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = [
  "backoff.*",
  "opentracing.*",
]
ignore_missing_imports = true

[tool.ruff.format]
quote-style = "single"
